refactor plugin structure
