--- plugins/notifications/styles.js
@@ -1,15 +1,0 @@
-'use strict';
-
-const red = '#da2100';
-const green = '#159600';
-
-const styles = {
-  errorToast: {
-    backgroundColor: red
-  },
-  successToast: {
-    backgroundColor: green
-  }
-};
-
-module.exports = styles;
