+++ src/views/appbar.js
@@ -1,0 +1,15 @@
+'use strict';
+
+const React = require('react');
+
+const TopBar = require('../components/top-bar');
+
+class AppbarView extends React.Component {
+  render(){
+    return (
+      <TopBar />
+    );
+  }
+}
+
+module.exports = AppbarView;
